+++ actix-web-lab/src/channel_body.rs
@@ -1,5 +1,4 @@
 use std::{
-    convert::Infallible,
     pin::Pin,
     task::{Context, Poll},
 };
@@ -8,14 +7,18 @@ use actix_web::body::{BodySize, MessageBody};
 use bytes::Bytes;
 use tokio::sync::mpsc::{error::SendError, UnboundedReceiver, UnboundedSender};
 
+use crate::BoxError;
+
 /// Returns a sender half and a receiver half that can be used as a body type.
 ///
 /// # Examples
 /// ```
 /// # use actix_web::{HttpResponse, web};
+/// use std::convert::Infallible;
 /// use actix_web_lab::body;
+///
 /// # async fn index() {
-/// let (mut body_tx, body) = body::channel();
+/// let (mut body_tx, body) = body::channel::<Infallible>();
 ///
 /// web::block(move || {
 ///     body_tx.send(web::Bytes::from_static(b"body from another thread")).unwrap();
@@ -26,41 +29,65 @@ use tokio::sync::mpsc::{error::SendError, UnboundedReceiver, UnboundedSender};
 /// HttpResponse::Ok().body(body)
 /// # ;}
 /// ```
-pub fn channel() -> (Sender, impl MessageBody) {
+pub fn channel<E: Into<BoxError>>() -> (Sender<E>, impl MessageBody) {
     let (tx, rx) = tokio::sync::mpsc::unbounded_channel();
     (Sender::new(tx), Receiver::new(rx))
 }
 
 /// A channel-like sender for body chunks.
 #[derive(Debug, Clone)]
-pub struct Sender {
-    tx: UnboundedSender<Bytes>,
+pub struct Sender<E> {
+    tx: UnboundedSender<Result<Bytes, E>>,
 }
 
-impl Sender {
-    fn new(tx: UnboundedSender<Bytes>) -> Self {
+impl<E> Sender<E> {
+    fn new(tx: UnboundedSender<Result<Bytes, E>>) -> Self {
         Self { tx }
     }
 
-    /// Submit a chunk of bytes to the response body stream.
-    pub fn send(&mut self, chunk: Bytes) -> Result<(), SendError<Bytes>> {
-        self.tx.send(chunk)
+    /// Submits a chunk of bytes to the response body stream.
+    ///
+    /// # Errors
+    /// Errors if other side of channel body was dropped, returning `chunk`.
+    pub fn send(&mut self, chunk: Bytes) -> Result<(), Bytes> {
+        self.tx.send(Ok(chunk)).map_err(|SendError(err)| match err {
+            Ok(chunk) => chunk,
+            Err(_) => unreachable!(),
+        })
+    }
+
+    /// Closes the stream, optionally sending an error.
+    ///
+    /// # Errors
+    /// Errors if closing with error and other side of channel body was dropped, returning `error`.
+    pub fn close(self, error: Option<E>) -> Result<(), E> {
+        if let Some(err) = error {
+            return self.tx.send(Err(err)).map_err(|SendError(err)| match err {
+                Ok(_) => unreachable!(),
+                Err(err) => err,
+            });
+        }
+
+        Ok(())
     }
 }
 
 #[derive(Debug)]
-struct Receiver {
-    rx: UnboundedReceiver<Bytes>,
+struct Receiver<E> {
+    rx: UnboundedReceiver<Result<Bytes, E>>,
 }
 
-impl Receiver {
-    fn new(rx: UnboundedReceiver<Bytes>) -> Self {
+impl<E> Receiver<E> {
+    fn new(rx: UnboundedReceiver<Result<Bytes, E>>) -> Self {
         Self { rx }
     }
 }
 
-impl MessageBody for Receiver {
-    type Error = Infallible;
+impl<E> MessageBody for Receiver<E>
+where
+    E: Into<BoxError>,
+{
+    type Error = E;
 
     fn size(&self) -> BodySize {
         BodySize::Stream
@@ -70,14 +97,16 @@ impl MessageBody for Receiver {
         mut self: Pin<&mut Self>,
         cx: &mut Context<'_>,
     ) -> Poll<Option<Result<Bytes, Self::Error>>> {
-        self.rx.poll_recv(cx).map(|opt_bytes| opt_bytes.map(Ok))
+        self.rx.poll_recv(cx)
     }
 }
 
 #[cfg(test)]
 mod tests {
+    use std::io;
+
     use super::*;
 
-    static_assertions::assert_impl_all!(Sender: Send, Sync);
-    static_assertions::assert_impl_all!(Receiver: Send, Sync, MessageBody);
+    static_assertions::assert_impl_all!(Sender<io::Error>: Send, Sync, Unpin);
+    static_assertions::assert_impl_all!(Receiver<io::Error>: Send, Sync, Unpin, MessageBody);
 }